move to vite as build tool & pnpm as package manager
