+++ package.json
@@ -5,54 +5,52 @@
   "author": "Bequest, Inc.",
   "private": false,
   "scripts": {
-    "build": "babel src --out-dir lib --copy-files && cp ./package.json lib/",
-    "dev": "babel-node src/index.js",
-    "start": "NODE_ENV=production node lib/index.js",
+    "preinstall": "npx only-allow pnpm",
+    "build": "vite build",
     "lint": "eslint ./src/",
-    "fix": "eslint ./src/ --fix",
+    "lint:fix": "eslint ./src/ --fix",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
   "devDependencies": {
-    "@babel/cli": "^7.13.16",
-    "@babel/core": "^7.14.0",
-    "@babel/eslint-parser": "^7.13.14",
-    "@babel/node": "^7.13.13",
-    "@babel/preset-env": "^7.14.0",
-    "@babel/preset-react": "^7.13.13",
-    "@material-ui/core": "^5.0.0-beta.5",
-    "@mui/icons-material": "^5.0.1",
-    "@mui/material": "^5.6.2",
-    "@storybook/addon-actions": "^6.4.1",
-    "@storybook/addon-docs": "^6.3.8",
-    "@storybook/addon-essentials": "^6.4.1",
-    "@storybook/addon-links": "^6.4.1",
-    "@storybook/react": "^6.4.1",
-    "babel-loader": "^8.2.3",
-    "babel-plugin-module-resolver": "^4.1.0",
-    "eslint": "^7.25.0",
+    "@babel/core": "^7.20.12",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.5",
+    "@storybook/addon-actions": "^6.5.15",
+    "@storybook/addon-docs": "^6.5.15",
+    "@storybook/addon-essentials": "^6.5.15",
+    "@storybook/addon-links": "^6.5.15",
+    "@storybook/react": "^6.5.15",
+    "@vitejs/plugin-react": "^3.0.1",
+    "babel-plugin-module-resolver": "^5.0.0",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-babel-module": "^5.3.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "prettier": "^2.2.1",
-    "pretty-quick": "^3.1.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-import-resolver-babel-module": "^5.3.2",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^3.4.1",
+    "eslint-plugin-react": "^7.32.1",
+    "prettier": "^2.8.3",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "vite": "^3.1.0"
   },
   "peerDependencies": {
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@mui/icons-material": "^5.0.1",
-    "@mui/material": "^5.6.2",
+    "@mui/material": "^5.11.5",
     "react": "^18.0.0",
     "react-dom": "^18.0.0"
   },
   "dependencies": {
-    "@emotion/react": "^11.8.1",
-    "@emotion/styled": "^11.8.1",
     "deepmerge": "^4.2.2",
-    "prop-types": "^15.7.2",
-    "tss-react": "^3.4.2"
+    "tss-react": "^4.5.2"
   }
 }
