+++ src/index.js
@@ -1,4 +1,3 @@
-export { makeStyles, withStyles, theme } from './theme';
-export { default as Spacer } from './Spacer';
-export { default as Modal } from './Modal';
-export { default as Table } from './table';
+export { makeStyles, withStyles, theme } from './theme.jsx';
+export { default as Spacer } from './components/Spacer.jsx';
+export { default as Modal } from './components/Modal.jsx';
