+++ src/stories/Modal.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Modal from '../Modal';
+import Modal from '../components/Modal.jsx';
 
 export default {
   title: 'Modal',
